fix entities jittering in water
Entities would jitter in water when Jesus was enabled. This is because of conflicting bounding box calculations between the server and client. This has been fixed by forcing the world to use a different bounding box calculation function (`BlockFluid.getCollidingBoundingBoxesPlayer` instead of `Block.getCollidingBoundingBoxes`) if the block is a fluid and the entity is a player.